+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Performance issue on v2 Flaw endpoint with include/exclude queryparams.
 - Added lost indexes for AffectsV1 and tests to ensure they are not deleted (OSIDB-4717).
 - Allow updates for flaw labels in any workflow state (OSIDB-4737)
+- Fix empty CVSS scores existing for not affected affects (OSIDB-4726)
 
 ## Removed
 - Remove CWE ID collection from collectors (OSIDB-4742)